requirements.txt
#VNPY required packages
attrs==25.4.0
backtrader==1.9.78.123
beautifulsoup4==4.14.3
Bottleneck==1.6.0
brotli==1.2.0
bs4==0.0.2
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.4
curl_cffi==0.13.0
deap==1.4.3
empyrical-reloaded==0.5.9
fastjsonschema==2.21.2
frozendict==2.4.7
greenlet==3.3.1
idna==3.11
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
loguru==0.7.3
lxml==6.0.2
msgpack==1.1.2
multitasking==0.0.11
mysql-connector-python==8.0.33
narwhals==2.15.0
nbformat==5.10.4
numpy==2.4.1
orjson==3.11.6
packaging==26.0
pandas==2.3.3
pandas-datareader==0.10.0
peewee==3.19.0
platformdirs==4.5.1
plotly==6.5.2
protobuf==3.20.3
pycparser==3.0
PyMySQL==1.1.2
pyproject_hooks==1.2.0
pyqtgraph==0.14.0
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
python-dateutil==2.9.0.post0
python-rapidjson==1.23
pytz==2025.2
pyzmq==27.1.0
QDarkStyle==3.2.3
QtPy==2.4.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rqdatac==3.4.1
scipy==1.17.0
shiboken6==6.8.2.1
simplejson==3.20.2
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.20
TA-Lib==0.6.8
tqdm==4.67.1
traitlets==5.14.3
tushare==1.4.24
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vnpy==4.3.0
vnpy_ctabacktester==1.3.0
vnpy_ctastrategy==1.4.0
vnpy_datamanager==1.2.0
vnpy_mysql==1.1.1
vnpy_sqlite==1.1.3
vnpy_tushare==1.4.21.0
websocket-client==1.9.0
websockets==16.0
yfinance==1.1.0
zstandard==0.25.0
# TraderMate API additional requirements
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
PyJWT>=2.8.0
redis>=5.0.0
rq>=1.15.0

# File-sync helper
watchdog>=2.1.0
